+++ mindsql/llms/__init__.py
@@ -1,6 +1,6 @@
+from .illm import ILlm
 from .anthropic import AnthropicAi
 from .googlegenai import GoogleGenAi
 from .huggingface import HuggingFace
-from .illm import ILlm
 from .llama import LlamaCpp
 from .open_ai import OpenAi
